+++ crates/glaredb/Cargo.toml
@@ -15,6 +15,7 @@ telemetry = {path = "../telemetry"}
 datafusion_ext = {path = "../datafusion_ext"}
 datafusion = {workspace = true}
 pgsrv = {path = "../pgsrv"}
+proxyutil = {path = "../proxyutil"}
 rpcsrv = {path = "../rpcsrv"}
 pgrepr = {path = "../pgrepr"}
 protogen = {path = "../protogen"}
