+++ crates/pgsrv/Cargo.toml
@@ -7,6 +7,7 @@ edition = {workspace = true}
 
 [dependencies]
 sqlexec = {path = "../sqlexec"}
+proxyutil = {path = "../proxyutil"}
 serde = { version = "1.0", features = ["derive"] }
 bytesutil = {path = "../bytesutil"}
 pgrepr = {path = "../pgrepr"}
