+++ crates/proxyutil/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "proxyutil"
+version.workspace = true
+edition.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+thiserror.workspace = true
+tracing = "0.1"
+futures = "0.3.28"
+async-trait = "0.1.72"
+reqwest = { version = "0.11.18", default-features = false, features = ["json", "rustls-tls"] }
+uuid = { version = "1.4.1", features = ["v4", "fast-rng", "macro-diagnostics"] }
+serde_json = "1.0.104"
