+++ crates/proxyutil/src/lib.rs
@@ -1,0 +1,2 @@
+//! Shared utilities for proxy services.
+pub mod cloudauth;
