+++ crates/rpcsrv/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod errors;
 pub mod handler;
+pub mod proxy;
 
 mod session;
